requirements.txt
asgiref==3.8.1
astroid==3.3.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssbeautifier==1.15.1
defusedxml==0.8.0rc2
dill==0.3.9
Django==5.1.1
django-allauth==65.0.2
djlint==1.35.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.10
isort==5.13.2
jsbeautifier==1.15.1
json5==0.9.25
mccabe==0.7.0
oauthlib==3.2.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
pylint==3.3.1
python-decouple==3.8
python-dotenv==1.0.1
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.1
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
